Medication + MedicationAdministration GET POST PUT DELETE

+++ resq_backend/src/main/java/at/resq/resq_backend/accidentPatient/medication/type/MedicationRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Time: 13:51
  */
 
-public interface MedicationRepository extends JpaRepository<Medication, Integer> {
+public interface MedicationRepository extends JpaRepository<Medication, Long> {
 }
